# ⚡ perf: Optimize BN254 pairing with Fuentes et al. hard exponentiation