## Summary

Implemented Fuentes et al. (2011) Algorithm  for the hard part of final exponentiation in BN254 pairing operations, replacing the previous naive implementation.

## Performance Results
- **Final exponentiation**: ~11ms → ~3ms (**~73% improvement**)
- **Pairing operations**: ~13ms → ~5ms (**~60% improvement**)

## Implementation Details
- Applied optimized sequence from Fuentes et al. 2011 as described in [this paper](https://eprint.iacr.org/2015/192.pdf) (algorithm 6)
- Reduces expensive field operations in Fp12 
- Results in a power of the optimal ate pairing, but maintains bilinearity which is all we need
- All existing tests pass

## Files Modified
- `src/crypto/bn254/pairing.zig` - Updated `final_exponentiation_hard_part()`
- `src/crypto/bn254/curve_parameters.zig` - Added required constants
- `src/crypto/bn254/Fp12Mont.zig` - Supporting operations

This optimization significantly improves performance for BN254 precompile operations

Co-authored with Sonnet 4